Only expose  WhiteSpaceTokenizer in tests
